Refactor extension slash commands into own file
- Weird circle imports again with the slash command classes
